+++ zenlang/src/ast/while_stmt.rs
@@ -38,7 +38,7 @@ impl Compile for AstWhileStmt {
         {
             let module = compiler.get_module();
             let body_index = module.opcodes.len() + 2;
-            module.opcodes.push(Opcode::BranchTrue(body_index as u32));
+            module.opcodes.push(Opcode::BranchTrue(body_index));
             br_exit_opcode_index = module.opcodes.len();
             module.opcodes.push(Opcode::Branch(0));
         }
@@ -53,10 +53,10 @@ impl Compile for AstWhileStmt {
             let module = compiler.get_module();
             exit_addr = module.opcodes.len() + 1;
             if let Opcode::Branch(addr) = &mut module.opcodes[br_exit_opcode_index] {
-                *addr = exit_addr as u32;
+                *addr = exit_addr;
             }
 
-            module.opcodes.push(Opcode::Branch(cmp_addr as u32));
+            module.opcodes.push(Opcode::Branch(cmp_addr));
         }
 
         // break statementss
@@ -70,7 +70,7 @@ impl Compile for AstWhileStmt {
                 let module = compiler.get_module();
                 for index in last.iter() {
                     if let Opcode::Branch(addr) = &mut module.opcodes[*index] {
-                        *addr = exit_addr as u32;
+                        *addr = exit_addr;
                     }
                 }
             }
@@ -93,7 +93,7 @@ impl Compile for AstWhileStmt {
                 let module = compiler.get_module();
                 for index in last.iter() {
                     if let Opcode::Branch(addr) = &mut module.opcodes[*index] {
-                        *addr = cmp_addr as u32;
+                        *addr = cmp_addr;
                     }
                 }
             }
